Make RAA-tools codebase for KMB compatible with BatchUploadTools
A first stab at making the old codebase compatible. The majority
of kmb_massload (which fetches and processes the metadata) comes
verbatim from RAA-tools and is in a terrible shape. But as the
code works (produces the expected output) cleanup and refactoring
is left to later smaller patches. Similarly a lot of functionality
will likely migrate from kmb_massload to KMBItem.

Task: T164566
Task: T164577