Use only primary classes for categorisation if matched
Provided a list of primary classes check first if any of the
item_class tags are in this list and can be used to generate a
category. If not then use the remaining to generate categories.

This also splits the handling of item_class tags and
item_keyword tags with all of the latter always getting used.

Task: [T160054](https://phabricator.wikimedia.org/T160054)